ajout de (int) pour abs dans IsCtrlPoint
git-svn-id: https://svn.code.sf.net/p/wired/code/trunk@154 6f5cd6f6-390f-0410-abd1-b47fe60d5253

+++ wired/src/plugins/akai_sampler/ASEnvel.cpp
@@ -31,7 +31,7 @@ ASEnvelSeg::~ASEnvelSeg()
 int ASEnvelSeg::IsCtrlPoint(wxPoint p, int precision)
 {
   for (unsigned int i = 0; i < nbpts; i++)
-    if ((abs(p.x - ratiox * points[i].x) <= precision) && (abs(p.y - ratioy * points[i].y) <= precision))
+    if ((abs((int)(p.x - ratiox * points[i].x) <= precision)) && (abs((int)(p.y - ratioy * points[i].y)) <= precision))
       return i;
   return -1;
 }
